feat(auth): unificar login docente+estudiante en / y arreglar bug de puerta equivocada
El selector "Selecciona tu perfil" (3 tarjetas → /teacher/login, /student/login,
/admin/login) era cosmético: la auth Supabase es agnóstica al rol y el backend lo
deriva de las memberships vía /api/auth/me. Quedarse en /teacher/login con cuenta de
estudiante dejaba al usuario parado en el formulario (GuestOnlyRoute solo redirige si
el rol coincide) mientras el header global ya mostraba su nombre.

- AppLanding (/) pasa de selector de 3 perfiles a login unificado (correo+contraseña).
  RootRedirect ya enruta a los autenticados por primary_role → el estado de "puerta
  equivocada" deja de existir por construcción. Microsoft OAuth queda tras el flag
  isMicrosoftLoginEnabled() (oculto hoy, handler intacto). Se preserva el flujo de
  inscripción por curso (activationContext → /auth/callback).
- Admin queda como acceso aparte y discreto (/admin/login, sin cambios).
- Se eliminan las rutas /teacher/login y /student/login (el catch-all * → / cubre
  links viejos); RequireRole y el enlace de StudentJoinPage repuntan a /.
- Se borran TeacherLoginPage y StudentLoginPage (+ tests); se reescribe AppLanding.test
  y se actualizan App/GuestOnlyRoute/StudentJoinPage tests. Runbook canónico al día.
