+++ frontend/src/app/App.test.tsx
@@ -8,9 +8,6 @@ vi.mock("@/app/auth/useAuth");
 vi.mock("@/features/teacher-authoring/TeacherAuthoringPage", () => ({
     TeacherAuthoringPage: () => <div data-testid="teacher-authoring-page">Teacher area</div>,
 }));
-vi.mock("@/features/teacher-auth/TeacherLoginPage", () => ({
-    TeacherLoginPage: () => <div data-testid="teacher-login-page">Teacher login</div>,
-}));
 vi.mock("@/features/teacher-dashboard/TeacherDashboardPage", () => ({
     TeacherDashboardPage: () => <div data-testid="teacher-dashboard-page">Teacher dashboard</div>,
 }));
@@ -40,9 +37,6 @@ vi.mock("@/features/auth-callback/AuthCallbackPage", () => ({
 }));
 vi.mock("@/features/student-auth/StudentJoinPage", () => ({
     StudentJoinPage: () => <div data-testid="student-join-page">Student join</div>,
-}));
-vi.mock("@/features/student-auth/StudentLoginPage", () => ({
-    StudentLoginPage: () => <div data-testid="student-login-page">Student login</div>,
 }));
 vi.mock("@/features/student-dashboard/StudentDashboardPage", () => ({
     StudentDashboardPage: () => (
@@ -163,21 +157,21 @@ describe("App admin shell layout", () => {
         expect(await screen.findByTestId("admin-login-page")).toBeTruthy();
     });
 
-    it("redirects an anonymous user to the teacher login route from /teacher/dashboard", async () => {
+    it("redirects an anonymous user from /teacher/dashboard to the unified login", async () => {
         vi.mocked(useAuth).mockReturnValue(baseContext);
 
         renderWithProviders(<App />, {
             initialEntries: ["/teacher/dashboard"],
         });
 
-        expect(await screen.findByTestId("teacher-login-page")).toBeTruthy();
+        expect(await screen.findByTestId("app-landing")).toBeTruthy();
     });
 
-    it("keeps the global SiteHeader on non-admin-dashboard routes", () => {
+    it("keeps the global SiteHeader on non-shell, non-landing routes", () => {
         vi.mocked(useAuth).mockReturnValue(baseContext);
 
         renderWithProviders(<App />, {
-            initialEntries: ["/teacher/login"],
+            initialEntries: ["/join"],
         });
 
         expect(screen.getByTestId("site-header")).toBeTruthy();
@@ -336,14 +330,24 @@ describe("App admin shell layout", () => {
         expect(await screen.findByTestId("student-join-page")).toBeTruthy();
     });
 
-    it("resolves the lazy student login route with an access token", async () => {
+    it("redirects the legacy /student/login path to the unified login", async () => {
         vi.mocked(useAuth).mockReturnValue(baseContext);
 
         renderWithProviders(<App />, {
-            initialEntries: ["/student/login?course_access_token=test-token"],
+            initialEntries: ["/student/login"],
         });
 
-        expect(await screen.findByTestId("student-login-page")).toBeTruthy();
+        expect(await screen.findByTestId("app-landing")).toBeTruthy();
+    });
+
+    it("redirects the legacy /teacher/login path to the unified login", async () => {
+        vi.mocked(useAuth).mockReturnValue(baseContext);
+
+        renderWithProviders(<App />, {
+            initialEntries: ["/teacher/login"],
+        });
+
+        expect(await screen.findByTestId("app-landing")).toBeTruthy();
     });
 
     it("does not render the global SiteHeader on /student/dashboard", async () => {
